requirements.txt
inflect
pandas
praw
python-dateutil
pyyaml
simplejson
tabulate
unidecode